Created test_pages.py
